// src/app/components/shared/models/order.ts
export class Order {
  id: number;
  processedOrder: boolean;
  allowShipping: boolean;
  sendAllProduct: boolean;
  orderNumber?: string;
  idSeller: number;
  nameSeller?: string;
  nitSeller?: string;
  idChannel: number;
  channel?: string;
  dateOrder: string;
  idStatusOrder: number;
  statusOrder?: string;
  costTotalOrder: number;
  costTotalShipping: number;
  commission: number;
  dateMaxDeliveryOrder: string;
  typeDespatchOrder?: string;
  identificationCard?: string;
  clientName?: string;
  clientTelephone?: string;
  clientAddress?: string;
  products?: (ProductsEntity)[] | Array<ProductsEntity>;
}

/**
 * Información del cliente
 * @export
 * @class InfoClient
 */
export class InfoClient {
  identificationCard?: string;
  clientName?: string;
  clientTelephone?: string;
  clientAddress?: string;
}

/**
 * Información del producto
 * @export
 * @class ProductsEntity
 */
export class ProductsEntity {
  idDetailProduct: number;
  idPicking: number;
  id: number;
  nameProduct?: string;
  productId?: string;
  sku?: number;
  ean?: string;
  checkProductToSend: boolean;
  reference?: string;
  idStatusProduct: number;
  statusProduct: string;
  quantity: number;
  costProduct: number;
  costShippingProduct: number;
  commission: number;
  dateMaxDeliveryProduct: string;
  typeDespatchProduct?: string;
  tracking?: string;
  carrier?: string;
  dateDelivery: string;
  enviosExito: boolean;
  fulfillment: boolean;
}

/**
 * Información de la categoría
 * @export
 * @class CategoryList
 */
export class CategoryList {
  root: string;
  name: string;
  id: string;
  buttonId: string;
  count: number;
}
// ...
export class InformationToForm {
  reversionRequestStatusId: number;
}

/**
 * Entidad que se emplea en los formularios de las diferente tablas de la aplicación,
 * Permite indicar el titulo, el tipo de formulario y la información que se le desee pasar al formulario.
 * @export
 * @class SearchFormEntity
 */
export class SearchFormEntity {
  title: string;
  type_form: string;
  btn_title: string;
  title_for_search: string;
  information: InformationToForm;
}